+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.5.0 - 2016-04-29
+### Added
+- Refactoring ctor.
+- Populate resource metadata from get and assignment.
+- Dataset collection.
+- RSpec tests for dataset collection.
+
 ## 0.4.0 - 2016-04-27
 ### Added
 - Date normalisation and conversion to ISO8601.